requirements.txt
six>=1.12.0
ckantoolkit>=0.0.3
goodtables==1.5.1
-e git+https://github.com/ckan/ckanext-scheming.git#egg=ckanext-scheming
